tests/wpt/web-platform-tests/html/cross-origin-embedder-policy/credentialless/iframe-coep-require-corp.tentative.https.html

<meta name="variant" content="?1-4">
<meta name="variant" content="?5-9">
<meta name="variant" content="?9-last">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="/common/utils.js"></script>
<script src="./resources/common.js"></script>
<script src="./resources/dispatcher.js"></script>
<script src="./resources/iframeTest.js"></script>
<script src="/common/subset-tests.js"></script>
<script>

const parent_coep_require_corp = newWindow(coep_require_corp);

subsetTest(iframeTest, "COEP:require-corp embeds same-origin COEP:none",
  parent_coep_require_corp, same_origin, coep_none, EXPECT_BLOCK);
subsetTest(iframeTest, "COEP:require-corp embeds cross-origin COEP:none",
  parent_coep_require_corp, cross_origin, coep_none, EXPECT_BLOCK);
subsetTest(iframeTest, "COEP:require-corp embeds same-origin COEP:credentialless",
  parent_coep_require_corp, same_origin, coep_credentialless, EXPECT_LOAD);
subsetTest(iframeTest, "COEP:require-corp embeds cross-origin COEP:credentialless",
  parent_coep_require_corp, cross_origin, coep_credentialless, EXPECT_BLOCK);
subsetTest(iframeTest, "COEP:require-corp embeds same-origin COEP:require-corp",
  parent_coep_require_corp, same_origin, coep_require_corp, EXPECT_LOAD);
subsetTest(iframeTest, "COEP:require-corp embeds cross-origin COEP:require-corp",
  parent_coep_require_corp, cross_origin, coep_require_corp, EXPECT_BLOCK);

// Using CORP:cross-origin might unblock previously blocked iframes.
subsetTest(iframeTestCORP, "COEP:require-corp embeds same-origin COEP:none",
  parent_coep_require_corp, same_origin, coep_none, EXPECT_BLOCK);
subsetTest(iframeTestCORP, "COEP:require-corp embeds cross-origin COEP:none",
  parent_coep_require_corp, cross_origin, coep_none, EXPECT_BLOCK);
subsetTest(iframeTestCORP, "COEP:require-corp embeds same-origin COEP:credentialless",
  parent_coep_require_corp, same_origin, coep_credentialless, EXPECT_LOAD);
subsetTest(iframeTestCORP, "COEP:require-corp embeds cross-origin COEP:credentialless",
  parent_coep_require_corp, cross_origin, coep_credentialless, EXPECT_LOAD);
subsetTest(iframeTestCORP, "COEP:require-corp embeds same-origin COEP:require-corp",
  parent_coep_require_corp, same_origin, coep_require_corp, EXPECT_LOAD);
subsetTest(iframeTestCORP, "COEP:require-corp embeds cross-origin COEP:require-corp",
  parent_coep_require_corp, cross_origin, coep_require_corp, EXPECT_LOAD);

</script>
