LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/credentialless/iframe-coep-require-corp.tentative.https-expected.txt


FAIL COEP:require-corp embeds same-origin COEP:none assert_equals: expected "block" but got "load"
FAIL COEP:require-corp embeds cross-origin COEP:none assert_equals: expected "block" but got "load"
PASS COEP:require-corp embeds same-origin COEP:credentialless
FAIL COEP:require-corp embeds cross-origin COEP:credentialless assert_equals: expected "block" but got "load"
PASS COEP:require-corp embeds same-origin COEP:require-corp
FAIL COEP:require-corp embeds cross-origin COEP:require-corp assert_equals: expected "block" but got "load"
FAIL COEP:require-corp embeds same-origin COEP:none, CORP:cross-origin assert_equals: expected "block" but got "load"
PASS COEP:require-corp embeds cross-origin COEP:none, CORP:cross-origin
PASS COEP:require-corp embeds same-origin COEP:credentialless, CORP:cross-origin
PASS COEP:require-corp embeds cross-origin COEP:credentialless, CORP:cross-origin
PASS COEP:require-corp embeds same-origin COEP:require-corp, CORP:cross-origin
PASS COEP:require-corp embeds cross-origin COEP:require-corp, CORP:cross-origin
